AP-441 Implement JSON Schema and validation
This PR establishes the JSON schema for the payload to submit an assessmnent.

The JsonSchemaValidator class will validate the pauyload and report any errors - this
should be called from the relevant controller action.

TODO: The schema is currently located in /config/api/assessment_schema.json.  This should
proobably be noved, as we will want to make the schema avaiable for download by clients so
that they can validate their payload before submitting.