AP-727 Create an Assessment controller
This PR is the first part of a major refactor to replace uploading a large JSON payload in one go
with a series of calls to build up the assessment.  As a result, the workflow services that have been
written so far will take their data from ActiveRecord models rather than parts of the JSON payload.

Threfore, the tests for these services have been disabled for the time being, until such time as all
the models have been created and the services updated to take their values from the models.