Create Applicant for Means Calculation Service
Create Applicants model that belongs_to the Assessment model and each assessment has one applicant.
Create ApplicantCreationService that validates the applicant params passed in and if successful creates an applicant and returns true else false.
Create ApplicantsController to call the ApplicationCreationService and render a success or error response depending on the result.
Create Tests