AP-786 Add temporary nocov to dated_struct
The property_assessment_spec implicitly tested the `[]=(name, value)`
method of open_struct. Changes to that class mean the method is no
longer tested so the build fails on code coverage.

There are other specs that will test that code (eg self_employed_spec)
but those tests are currently disabled.

Temporarily add nocov tags to the method until those tests are enabled
or we are sure that the method can be removed.