Fix Faker usage
The update of Faker to 2.1.0 introduced a number of breaking changes.
This commit fixes the use of Faker so that all specs continue to pass.

The main change is the mandatory use of keyword arguments: for example,
Faker::Date.between(18.years.ago, 70.years.ago) now needs to be
Faker::Date.between(from: 18.years.ago, to: 70.years.ago) and
Faker::Number.decimal(4, 2) is now Faker::Number.decimal(l_digits:
4, r_digits: 2).