AP-998 Fix Property Calculation
The capital calculation is incorrect for properties where ownership is
shared with a housing association.

In these instances the housing association share is a percentage of the
market value rather than a percentage of the net equity.

This commit amends the calculate_net_equity method on the property model
so that the correct value is calculated dependent on whether a
housing association owns a share of the property.

A (disabled) spec already existed for this scenario. It has been enabled
and updated with correct values derived from the guidance in the Jira
story: https://dsdmoj.atlassian.net/wiki/spaces/ATPPB/pages/1339654172/3.+Detailed+Means+Calculation+-+Capital+Assets?preview=/1339654172/1547763935/Property.pdf

This commit also includes a small change to account for scenarios where
the main dwelling equity disregard is greater than the equity, to limit
assessed_equity to 0 instead of returning a negative value.