AP-997 Remove upper threshold
For cases where the applicant has an involvement type of 'applicant'
and the matter proceeding type is 'domestic_abuse', the upper capital
threshold does not apply. All applicants above the lower threshold
should have a result of contribution_required.

For MVP all cases are applicant/domestic_abuse, so to change this simply
amend the capital_summary model to remove the upper threshold.

It can be reinstated when other involvement types/matter types are
introduced to the service.