AP-1099 Add earned income end point
Creates a controller and route to handle earned income payloads (made up
of wages and benefits in kind).

Also adds an earned incomes creation service, which will create database
records based on the payload.

There are currently two tests which do not currently passed. These test
validation of the payload when a wage payment date is before the
assessment submission date, and when an empty wage array is passed. They
are currently excluded from the rspec suite to allow CI pipeline to
complete.