AP-1103 Calculate gross income threshold
CFE needs to be able to derive the correct value for the upper gross
income threshold and store it in the database so it can be used in
eligibility calculations.

This commit:

* creates a migration to add a field to gross_income_summary to hold the
upper threshold

* adds a new gross_income_collator class to derive the correct threshold
from config based on date and number of dependants and persist it to the
database

* adds the correct threshold data to config/thresholds files. Only the
data for 2019 is currently available. The data for 2018 and 2020 will
need to be updated when the correct values are known.