Set CircleCI to use Europe/London as Timezone
This fixes integration tests failures. Related to the Threshold model. CircleCI was using UTC/GMT, whilst
the app uses BST.

The value_for method in the Threshold model takes in a submission date of 06/04/2020 as the :at parameter so in the 'time < at'
it says is the BST value of 2020-04-06 00:00:00 +0100 before the 06/04/2020(which is the same as 2020-04-06 00:00:00 +0000 in circleci)
An hour difference causes the issue so setting the circleci to use the same timezone as we run tests "Europe/London" will remove this issue.