Update GrossIncomeSummaryDecorator and its test
Update GrossIncomeSummaryDecorator to include irregular_income into the json response

Only include monthly_student_loan key when student loan is not in the other income sources. This prevents confusion with monthly_student_loan
and the student_loan in the other income. Also, this will be updated in the future to no longer use student_loan in other income and thus adding the key monthly_student_loan
into the json response.

Update the decorator tests.