AP-1441 Add client id to integration tests
The integration test spreadsheet has been updated so that all
transactions in the other_incomes, outgoings and state_benefits
categories now include a client_id.

As a result, the create_client_id method in DeeplyNestedPayloadGenerator
is no longer necessary. This commit removes it and adjusts the logic so
that the client_id from the spreadsheet is used.