AP-1806 Fix flagging of childcare payments
Childcare payments are being flagged for amount variation/frequency
issues even if they have been disallowed from the calculation.

This commit changes the logic so that childcare is only flagged if the
payment is included in the calculation. It does this by adding extra
guards so that childcare payments are only checked by remark generators
when they have value > 0 (which indicates that the payment was included
in the calculation).