Fix bug that sends exception object to raven capture message
We capture api param errors with an array of string instead of an exception object for simplification, so that user
doesn't receive the error class in a response.

Have a seperate Raven.capture_exception when capturing param errors because render_unprocessable is used from child controllers
to signify an internal error but without providing error class info etc.