AP-1915 Run CFE on Alpine docker image
CFE currently runs on the Ruby 2.7.2 docker image which contains more
code than is needed. This commit switches to the smaller Alpine image,
also used by Apply, resulting in smaller image sizes.

As part of this change it is necessary to remove the platform constraint
from the tz-info gem as this prevents the gem installing on Alpine. This
also brings it in-line with the Apply gemfile.