AP-2805 Fix GrossIncomeCollator floating point error
Upgrading to Ruby 3.1.0 has caused spec/services/collators/gross_income_collator_spec.rb:56
to start failing due to a floating point comparison error. This seems
to be caused by the categorise_income method of the GrossIncomeCollator
class, which sums monthly_income amounts.

Although the test data for these monthly_income amounts are all 2dp,
summing them results in 171.57999999999998, which now results in a
failure when carrying out the comparison in the test.

To fix this, this commit amends the categorise_income to convert all
monthly_income amounts to BigDecimal before addition. This prevents
rounding errors and results in the correct value of 171.58.