AP-2805 Fix StateBenefitsCalculator floating point error
Upgrading to Ruby 3.1.0 has caused spec/services/calculators/state_benefits_calculator_spec.rb:40
to start failing due to a floating point comparison error.

That test expects the state benfit amount to equal 216.67 + 88.3 however
calculates that as 304.96999999999997 not 304.97 and now complains that
the amounts don't match...

To fix this, this commit hard codes the expected amount as 304.97 and so
the comparison is now always correct.