Move all endpoints to version 4
Version 3 is deprecated and should not really
be documented at all. CFE is to be changed to
make version 4 the default and this change is in
line with that.

Once the default is changed we could remove version
3 swagger documentation entirely although I would be
minded to retain it until it ceases to be available
at all in the App (i.e. when it is EoL).