AP-3297 Add thresholds to proceeding types and create Threshold Popualtor
Up until now, client involvement type has not been considered in CFE, because Apply only allowed one type: Applicant. Now that multiple client involvement types are allowed, it has a bearing on whether or not the upper thresholds are waived or not. In short, for Client involvement type "A" (applicant) upper thresholds are waived for Domestic Abuse cases only. In all other cases, they remain as specified by the thresholds.

In accordance with the principle of only one source for information, the Legal Framework API (LFA) is the source for this information, and so CFE needs to ask LFA whether or not to waive the thresholds for each proceeding type / client involvement type combination.

In a previous PR, a Proceeding Types record was created, and this is written when Apply requests an assessment with version 5, and sends proceeding type/client information data to the new /proceeding_types endpoint.

This change does the following to enable the calculations to take client involvement type into consideration:

adds three fields to the ProceedingType model:
capital_upper_threshold
gross_income_upper_threshold
disposable_income_upper_threshold
At the start of the main workflow, a new service is called for version 5 assessments which queries LFA with all the proceeding type/client involvement type combinations on the assessment, and then populates these three fields with eithet the standard upper thresholds or the normal upper thresholds as appropriate
later in the workflow, when the Eligibility records are created, for version 5 assessments, these thresholds are copied from the ProceedingType (for earlier versions, the existing method is still. used until versions 3 and 4 are deprecated)