AP-3108 Ensure code coverage at 100%
This PR addresses two issues:

1) The eager_loading setting in /config/test.rb was set to false.  This means that only those files
actually used during a test run are included in the coverage results.  By setting this to true, this
enxures that all files are included in the test results, and files that are never hit are shown as
having no coverage.

2) It fixes a flickering test error where employments were being created by the factory as
Job 1, Job 2 etc were being incorrectly sorted (Job 10 coming before Job 9) and therefore
not matching the expected order.  This fixes it by making sure the numerical element is a
four-difit leading zero number (e.g. Job 0010)