AP-3509 Gracefully prevent calling GET on the same assessment multiple times
Users reported getting exceptions when calling GET more than once for the same assessment.

This was because during the assessment process, eligibilty records are created and were
failing because they already existed.  In order to enable calling GET on the assessment to
be truly idempotent, we wold have had to delete any eligibilty records, remarks records and
zeroise all the totals on the summary records.

For that reason it was felt better to gracefully refuse to carry out a seconds assesment as
this would not be a normal use-case anyay, and return and explanatory message.

The swagger documentation has been updated to reflect this.