Add CFE intgration tests for disposable income summary
We should be using these CFE integration tests to cover overall
results more extensively, as a separate set of testing to the
Worksheet based integration tests. The latter are more intended
for use by the BA to setup scenarios.

This adds two tests, one to test existing disposable income
result for a prexexisting tests requests. This is to provide
a regression and comparison for the second test, which uses
the regular transaction mechanism to add income, outgoings
and husing benefits.

The second test uses the exact equivalent values of the
the first test's request but supplied via the regular
transaction's flow.