Refactor `#applicant_has_dependant_child?` method
An applicant can only be eligible for childcare costs if they have a dependant
child.

This refactors a private method in the `Collators::ChildcareCollator` to be
clearer and reveal the intent more explicitly.

A new date method `becomes_adult_on` has been added to the `Dependant` model,
which returns the date the individual turns 16 years of age.