Define Faraday gem as an explicit dependency
Throughout the application, we are using the [Faraday gem] to make
requests to the Legal Framework API.

Until now, we have been relying on Faraday being pulled in by another of our
dependencies.

Ideally we would be using [Net::HTTP] since it comes as part of Ruby, but
at the very least we should be explicit about the dependencies our application
needs.

[Faraday gem]: https://github.com/lostisland/faraday
[Net::HTTP]: https://rubyapi.org/3.1/o/net/http