Fix employment without payments bug
An [existing bug resurfaced recently][sentry].

When an assessment has an employment but the employment has no employment
payments, a `NoMethodError` is raised when calculating the range of monthly
employment income payments in `Utilities::EmploymentIncomeVariationChecker`.

While this scenario is unlikely, it is possible -so we should account for it.

In the instance linked above, the applicant specified that they were not
employed, but the HMRC response indicated that the applicant had an employment
(albeit they had no employment payments).

This change ensures the most recent employment payment is only used for monthly
employment values when there is at least one employment payment, and the range
of employment income payments is below the threshold.

If there are no employment payments or the range exceeds the threshold, the
blunt average of all employment payments is used. The blunt average will return
`0` if there are no employment payments.

As part of this work, the `#calculate!` method has been simplified so employment
records are only update once.

The associated specs have also been refactored, with the intention of testing
the behaviour of the `#calculate!` method rather than focusing on implementation
details.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3587)

[sentry]: https://sentry.io/organizations/ministryofjustice/issues/3345805354