Fix: remove test flakiness
In tests "aged 15" dependents had a date of birth set to a range that included 16 years before the submission date, meaning their 16th birthday was on the submission date, meaning for the purposes of the calculation they were 16. This was causing approximately 1 test run in every 366 to fail.

This PR fixes the off-by-one error, and also adjusts the other age traits to use all valid date of birth values for those ages.

Thanks to the EFE team