Update dependant allowances
From 10th April 2023, changes to means assessment fixed allowances for
dependants will go live.

This updates the dependant allowances to the new amounts. All other thresholds
have been left unchanged.

There is an allowance for dependant partners, but this hasn't been captured
since we expect to integrate with the new `CFE-Civil` service before releasing
partner means functionality in Apply.

These allowances should be applied for applications with submission dates on or
after 2023-04-10.

Apply handles the logic surrounding application submission dates and delegated
functions.

The Apply service passes the submission date when making requests to the
assessment endpoint, so we do not need to worry about calculating it for an
application.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3950)